# Implement FastLock for DatabaseHeader