… core/storage/pager.rs
… core/translate/mod.rs
… core/translate/pragma.rs
… core/vdbe/builder.rs
… core/vdbe/mod.rs